requirements.txt
flask==1.1.2
python-dotenv==0.14.0

pytest==5.4.3
